requirements.txt
adjustText==0.7.3
aenum==3.1.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appdirs==1.4.4
appnope==0.1.0
appscript==1.0.1
APScheduler==3.8.1
asn1crypto==1.0.1
astroid==2.3.1
astropy==3.2.2
atomicwrites==1.3.0
attrs==19.2.0
autograd @ git+https://github.com/HIPS/autograd@01eacff7a4f12e6f7aebde7c4cb4c1c2633f217d
azure-core==1.19.0
azure-storage-blob==12.9.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
bs4==0.0.1
case-conversion==2.1.0
cdata-microsoftplanner-connector==21.0.7930
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
clipboard==0.0.4
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
colour==0.1.5
conda==4.7.12
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0
cryptography==35.0.0
cursor==1.3.4
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.4
distributed==2.5.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
fsspec==0.5.2
future==0.17.1
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
greenlet==0.4.15
h5py==2.9.0
HeapDict==1.0.1
html-text==0.5.2
html5lib==1.0.1
idna==2.8
imageio==2.6.0
imageio-ffmpeg==0.4.5
imagesize==1.1.0
importlib-metadata==0.23
infi.systray==0.1.12
inflect==5.3.0
ipykernel==5.1.2
ipython==7.8.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jhwutils==1.0
Jinja2==2.10.3
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
maybe-else==0.2.1
mbstrdecoder==1.1.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
moviepy==1.0.3
mpmath==1.1.0
msal==1.15.0
msgpack==0.6.1
msoffcrypto-tool==4.12.0
msrest==0.6.21
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
ntlm-auth==1.5.0
numba==0.45.1
numexpr==2.7.0
numpy==1.17.2
numpydoc==0.9.1
O365==2.0.16
oauthlib==3.1.1
Office365-REST-Python-Client==2.3.8
olefile==0.46
openpyxl==3.0.0
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parsedatetime==2.6
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
pathmagic==0.3.14
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
prettierfier==1.0.3
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydub==0.25.1
pyflakes==2.1.1
Pygments==2.4.2
pyinstrument==4.0.4
pyiotools==0.3.18
PyJWT==2.3.0
pylint==2.4.2
pymiscutils==0.3.14
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyPDF2==1.26.0
pyperclip==1.8.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.15.4
PySocks==1.7.1
pysubtypes==0.3.18
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.0
python-docx==0.8.11
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
readchar==3.0.4
regex==2021.10.23
requests==2.22.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rope==0.14.0
ruamel_yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
SharePlum==0.5.1
simplegeneric==0.8.1
simplejson==3.17.5
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
statsmodels==0.10.1
stringcase==1.2.0
sympy==1.4
tables==3.5.2
tabulate==0.8.9
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typepy==1.3.0
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
